Updated buffer sizes to reflect actual use
There was potential for buffer issues to inhibit correct communication. Some of the data was longer than the parameter buffer accounted for. The parameter buffer length has been increased.
There was also a potential for the send buffer to become saturated too easily. The send buffer has been increased (this also increased the receive buffer).