LPF Completion
Appears to be stable for the time being. The LPF is read and allocation resets are accomplished whether power is on or off. Changes primarily to the Allocation reset check function and the I2C read power failure times function. Finally, added a get function for the power failure failure and restoration times.