Bugfix - Relay Control
The relay had only 'off' and 'on' options which was confusing. Was 'off' supposed to turn the relay off? Was 'on' supposed to keep it on or work in 'auto' mode.

Now there are three options which provide better and more sensical control. Off, On, Auto. Just as the names state it works.